requirements.txt
robotframework==3.1.2
robotframework-seleniumlibrary==4.3.0
anytree==2.8.0
pyyaml==5.3.1
robotframework-lint==1.0
click==7.1.1
